docker: Warn about broken host setup based on aufs
There is a nasty problem with legacy aufs: wic tries to find out the
block size of the filesystem that holds the partition images, but aufs
does not seem to implement this properly, returning 0 at least on Debian
Jessie. That makes wic become upset and through a division-by-zero
exception soon after.

Catch this case by warning the user about the inappropriate docker
setup during container start.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>