More details on docker usage
I would suggest to get more details in the doku on how to use the docker container.

I learned my lessons on how to use it and this is what works perfect and comfortable. 

Open for comments on this.

The two links https://raw.githubusercontent.com/siemens/kas/master/docs/userguide.rst should be cross references to the main userguide.rst. Didn't know yet how to do it with *.rst files