kas-docker: Podman: Fixing isar builds failing with exec format errors
When trying to run kas-docker with --isar on podman the build fails
with some "exec format" errors when the target has a different
architecture than the host. The reason is that /proc/sys/fs/binfmt_misc
is not mounted as expected inside the container.

The way the /proc filesystem is mounted seems to differ between podman
and docker. See [1] for the podman issue tracker. This is a temporary
fix and should be removed until the upstream issue was resolved.

[1] https://github.com/containers/podman/issues/4725

Signed-off-by: Florian Bezdeka <florian@bezdeka.de>