kas: allow toggling the bitbake "-k" argument
Not everyone might want to continue for a long time after the first taks
failed. That seems to be a good choice for interactive local build, but
a bad default choice for CI.

In fact it is bad that is not the bitbake default and set
unconditionally. But that is what we had, so to not break existing
setups just allow to turn it off.

Signed-off-by: Henning Schild <henning.schild@siemens.com>